Updated deploy script to pass correct parameters and permit absence of registry

+++ openchain/example/chaincode/asset_manager/deploySecurityContract.sh
@@ -1,6 +1,12 @@
 #!/bin/bash
 set -e
 
+if [[ $# -lt 3 ]]; then
+	echo "ERROR: Insufficient arguments to script"
+	echo "Usage: $0 <chaincode user> <chaincode path> <security name> [<dev mode (0|1)>]"
+	exit
+fi
+
 echo "--> Deploy asset contract... "
 
 #prereqs/assumptions
@@ -15,7 +21,7 @@ echo "--> Deploy asset contract... "
 #	security identifier	
 #	dev mode (1=YES, 0=NO)
 #EXAMPLE:
-#	./deploySecurityContract.sh chrisP myChainCodeID github.com/openblockchain/obc-peer/openchain/example/chaincode/simpleFinancialSecurity/ "IBM" 1
+#	./deploySecurityContract.sh chrisP github.com/openblockchain/obc-peer/openchain/example/chaincode/simpleFinancialSecurity/ "IBM" 1
 
 
 HOST=0.0.0.0
@@ -27,16 +33,24 @@ chainCodePath=$2
 securityName=$3
 devMode=$4
 
+registryNameFile=registry.address
 deployMode="-p"
 deployName=$chainCodePath
 pid=-1
 
-registryName=`cat registry.address`
+if [[ ! -e $registryNameFile ]]; then
+	echo Registry chaincode is missing... Will proceeed without
+else
+	registryName=`cat $registryNameFile`
+fi
 
 #Login to the local OBC peer
 echo "Logging in $chainCodeUser..."
 cd $GOPATH/src/github.com/hyperledger-incubator/obc-peer
 ./obc-peer login $chainCodeUser 
+
+OPENCHAIN_PEER_ADDRESS=$HOST:$PORT
+export OPENCHAIN_PEER_ADDRESS
 
 directoryName=$GOPATH/src/$chainCodePath
 if [[ $devMode -eq 1 ]]; then
@@ -49,7 +63,7 @@ if [[ $devMode -eq 1 ]]; then
 	cd $directoryName
 	echo "Building GO file $fileName..."
 	go build
-	OPENCHAIN_CHAINCODE_ID_NAME=$securityName OPENCHAIN_PEER_ADDRESS=$HOST:$PORT $fileName &
+	OPENCHAIN_CHAINCODE_ID_NAME=$securityName $fileName &
 	pid=$!
 	deployMode="-n"
 	deployName=$securityName
@@ -58,23 +72,25 @@ fi
 #Deploy the chaincode
 cd $GOPATH/src/github.com/hyperledger-incubator/obc-peer
 echo "Deploying contract $securityName with constructor init..."
-constructorJSON+='{"Function":"init", "Args": ["'
-constructorJSON+=$securityName
-constructorJSON+='"]}'
+constructorJSON="{\"Function\":\"init\", \"Args\": [\"$chainCodeUser\", \"$securityName\"]}"
 echo "Constructor args: $constructorJSON"
 ./obc-peer chaincode deploy -u $chainCodeUser $deployMode $deployName --ctor="$constructorJSON" > $directoryName/$securityName
 chainCodeAddress=$(cat $directoryName/$securityName)
 rm $directoryName/$securityName
 
-#Capture the contract hash and register it in the asset registry
-echo "Contract name: $chainCodeAddress"
-invokeConstructor='{"Function": "register", "Args": ["'$securityName'", "'$chainCodeAddress'"]}'
-./obc-peer chaincode invoke -u $chainCodeUser -n $registryName -c "$invokeConstructor"
+if [[ $registryName -ne "" ]]; then
+	#Capture the contract hash and register it in the asset registry
+	echo "Contract name: $chainCodeAddress"
+	invokeConstructor='{"Function": "register", "Args": ["'$securityName'", "'$chainCodeAddress'"]}'
+	./obc-peer chaincode invoke -u $chainCodeUser -n $registryName -c "$invokeConstructor"
+fi
 
-#kill the local chain code that we started earlier if it is still running
-if [[ $pid -eq 0 ]] &&  ps -p $pid > /dev/null ;
-then
-	kill $pid
+if [[ $devMode -eq 1 ]]; then
+	#kill the local chain code that we started earlier if it is still running
+	if [[ $pid -eq 0 ]] &&  ps -p $pid > /dev/null ;
+	then
+		kill $pid
+	fi
 fi
 
 echo "Chaincode $securityName has been compiled, registered and deployed!"
